rm boston, add leukemia and libsvm
